# design-patterns/facade/shipping/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cloud Shipping API",
        "title": "Cloud Shipping API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/addresses": {
            "get": {
                "description": "get addresses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Get addresses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "create addresses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Submit addresses",
                "parameters": [
                    {
                        "description": "Create addresses",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.LocationDto"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
        "/addresses/:addressID": {
            "get": {
                "description": "get a address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Get a address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address ID",
                        "name": "addressID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
...
        "/pickups": {
            "get": {
                "description": "get carrier pickups",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Carrier Pickup"
                ],
                "summary": "Get carrier pickups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "update a pickup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Carrier Pickup"
                ],
                "summary": "Update a pickup",
                "parameters": [
                    {
                        "description": "Update a pickup",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PickUpDto"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
...
        "/shipments": {
            "get": {
                "description": "get shipments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shipment"
                ],
                "summary": "Get shipments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "create a shipment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shipment"
                ],
                "summary": "Submit a shipment",
                "parameters": [
                    {
                        "description": "Create a shipment",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ShippingDto"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Address": {
            "type": "object",
            "required": [
                "city",
                "country",
                "email",
                "first_name",
                "last_name",
                "phone_no",
                "state",
                "street_name",
                "street_no",
                "zip"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 50
                },
                "country": {
                    "maxLength": 50,
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Country"
                        }
                    ]
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 50
                },
                "phone_no": {
                    "type": "string",
                    "maxLength": 50
                },
                "province": {
                    "type": "string"
                },
                "state": {
                    "type": "string",
                    "maxLength": 50
                },
                "street_name": {
                    "type": "string",
                    "maxLength": 50
                },
                "street_no": {
                    "type": "string",
                    "maxLength": 50
                },
                "terminal_address_id": {
                    "type": "string"
                },
                "zip": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "domain.Country": {
            "type": "string",
            "enum": [
                "NG",
                "US",
                "UK",
                "UAE",
                "NG"
            ],
            "x-enum-varnames": [
                "NG",
                "US",
                "UK",
                "UAE",
                "NIGERIA"
            ]
        },
...
        "domain.LocationDto": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Address"
                    }
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
        "domain.PickUpDto": {
            "type": "object",
            "required": [
                "carrier_id",
                "pick_up_at",
                "shipping_id",
                "status"
            ],
            "properties": {
                "carrier_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "pick_up_at": {
                    "type": "string"
                },
                "shipping_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "domain.ProductType": {
            "type": "string",
            "enum": [
                "animal",
                "plant",
                "appareal",
                "book",
                "cosmetics",
                "electronics",
                "watery",
                "ammunition"
            ],
            "x-enum-varnames": [
                "Animal",
                "Plant",
                "Appareal",
                "Book",
                "Cosmetics",
                "Electronics",
                "Watery",
                "Ammunition"
            ]
        },
...
        "domain.Response": {
            "type": "object",
            "properties": {
                "result": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "domain.ShippingDto": {
            "type": "object",
            "required": [
                "carrier_id",
                "delivery_address",
                "description",
                "pick_up_address",
                "product_type"
            ],
            "properties": {
                "carrier_id": {
                    "type": "string"
                },
                "delivery_address": {
                    "$ref": "#/definitions/domain.Address"
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 6
                },
                "pick_up_address": {
                    "$ref": "#/definitions/domain.Address"
                },
                "product_type": {
                    "$ref": "#/definitions/domain.ProductType"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
    }
}